@@ -34,61 +34,63 @@
   "license": "MIT",
   "dependencies": {
     "animate.css": "3.4.0",
-    "autobahn": "0.9.8",
-    "babel": "5.8.23",
-    "babel-core": "5.8.33",
-    "bootstrap-sass": "3.3.5",
-    "d3": "3.5.10",
-    "dotenv": "1.2.0",
-    "font-awesome": "4.4.0",
+    "autobahn": "0.9.9",
+    "bootstrap-sass": "3.3.6",
+    "d3": "3.5.12",
+    "fixed-data-table": "0.6.0",
+    "font-awesome": "4.5.0",
     "history": "~1.13.1",
     "lodash": "3.10.1",
     "moment": "2.10.6",
     "numeral": "1.5.3",
     "nvd3": "1.8.1",
-    "react": "0.14.0",
+    "react": "0.14.3",
+    "react-dimensions": "1.0.1",
     "react-dom": "0.14.3",
-    "react-nvd3": "0.3.0",
+    "react-nvd3": "0.4.0",
     "react-popout": "0.4.0",
     "react-router": "~1.0.2",
-    "react-sparklines": "1.4.2",
-    "rx": "4.0.7",
-    "yargs": "3.18.0"
+    "react-sparklines": "~1.4.2",
+    "rx": "4.0.7"
   },
   "devDependencies": {
-    "autoprefixer": "6.0.3",
+    "autoprefixer": "6.2.1",
     "autoprefixer-loader": "3.1.0",
+    "babel": "5.8.23",
+    "babel-core": "5.8.33",
     "babel-eslint": "4.1.1",
     "babel-loader": "5.0.0",
     "babel-plugin-react-transform": "1.1.1",
     "babel-runtime": "5.8.20",
     "chalk": "1.1.1",
     "copy-webpack-plugin": "0.3.3",
-    "css-loader": "0.23.0",
-    "eslint": "1.0.0",
+    "css-loader": "0.23.1",
+    "dotenv": "1.2.0",
+    "eslint": "1.10.3",
     "eslint-config-airbnb": "2.1.1",
-    "eslint-plugin-react": "3.3.1",
-    "extract-text-webpack-plugin": "0.9.0",
-    "file-loader": "0.8.4",
-    "html-webpack-plugin": "1.6.1",
-    "isparta-loader": "1.0.0",
+    "eslint-plugin-react": "3.12.0",
+    "extract-text-webpack-plugin": "0.9.1",
+    "file-loader": "0.8.5",
+    "html-webpack-plugin": "1.7.0",
+    "isparta-loader": "2.0.0",
     "jasmine-core": "2.4.1",
     "json-loader": "0.5.4",
-    "karma": "0.13.8",
+    "karma": "0.13.15",
     "karma-chrome-launcher": "0.2.2",
-    "karma-coverage": "0.5.0",
+    "karma-coverage": "0.5.3",
     "karma-jasmine": "0.3.6",
-    "karma-spec-reporter": "0.0.22",
+    "karma-spec-reporter": "0.0.23",
     "karma-webpack": "1.7.0",
-    "node-sass": "3.3.3",
-    "react-addons-test-utils": "0.14.0",
+    "node-sass": "3.4.2",
+    "react-addons-test-utils": "0.14.3",
     "react-transform-catch-errors": "1.0.0",
-    "react-transform-hmr": "1.0.0",
-    "redbox-react": "1.0.4",
-    "sass-loader": "3.0.0",
+    "react-transform-hmr": "1.0.1",
+    "redbox-react": "1.2.0",
+    "sass-loader": "3.1.2",
     "style-loader": "0.13.0",
-    "url-loader": "0.5.6",
-    "webpack": "1.11.0",
-    "webpack-dev-server": "1.10.1"
+    "url-loader": "0.5.7",
+    "webpack": "1.12.9",
+    "webpack-dev-server": "1.14.0",
+    "yargs": "3.31.0"
   }
 }
